refactor dev example
